Added support for request using plugin name (#33)
Currently, if reqeusting from plugin, you had to specify the plugin unique name. This is irrelevant for one instance plugins like aggregator or execution. Now we can use the same api, but if core cant find the unique_name in his list, it will try to look for the plugin name and not unique name